% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.EEG.R
\name{concatenate.EEG}
\alias{concatenate.EEG}
\title{A function to concatenate EEG data}
\usage{
concatenate.EEG(path_FW, path_BW, name_subj, TF_del, path_save)
}
\arguments{
\item{path_FW}{Path to the forward .ep files to concatenate}

\item{path_BW}{Path to the backward .ep files to conacatenate}

\item{name_subj}{Character vector containing a name (usually subjects) per concatenated file. The length of this vector should match the numner of element present in path_FW and path_BW}

\item{TF_del}{Number of time frames to delete. The time frames will be half deleted from the FW and half deleted from the BW}

\item{path_save}{Path to the folder where the concatenated files chould be saved.}
}
\description{
This function allows to create .ep files (Cartool format) with the concatenation of EEGs. The function takes as input the path to a folder containing forward and backward files and concatenate them according to different parameters.
}
\keyword{EEG}
\keyword{concatenate}
