% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.eph.ep.folder.R
\name{convert.eph.ep.folder}
\alias{convert.eph.ep.folder}
\title{A function to convert an entire folder from .eph to .ep}
\usage{
convert.eph.ep.folder(path)
}
\arguments{
\item{path}{Path to the folder containing some .eph files}
}
\description{
This package uses .ep format. Use this function to convert a folder containing .eph files into .ep files
}
\keyword{.ep}
\keyword{.eph}
\keyword{convert}
\keyword{ep}
\keyword{folder}
