% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.eph.ep.R
\name{convert.eph.ep}
\alias{convert.eph.ep}
\title{A function to convert .eph files into .ep files}
\usage{
convert.eph.ep(file)
}
\arguments{
\item{file}{Path or dataframe containing the .eph file}
}
\description{
This package uses the simple .ep format. This function converts the .eph into .ep format
}
\keyword{.ep}
\keyword{.eph}
\keyword{convert}
\keyword{ep}
\keyword{eph}
