% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.tva.files.R
\name{create.tva.files}
\alias{create.tva.files}
\title{Function to create the Triggers Validation Files for Cartool}
\usage{
create.tva.files(data, subj.var, CR.var, RT.var, trig.var, path.save)
}
\arguments{
\item{data}{Name of the data frame containing the results}

\item{subj.var}{Name of the subject's ID variable (tidy format, i.e. variable name in the data frame without $ or "")}

\item{CR.var}{Name of the accuracy variable (tidy format)}

\item{RT.var}{Name of the RT variable (tidy format)}

\item{trig.var}{Name of the variable containing the triggers (tidy format)}

\item{path.save}{Path to the folder where to save the .tva files}
}
\description{
This functions allows the user to create automatically the .tva files based on the behavioral results dataframe
}
\keyword{.TVA}
\keyword{.tva}
\keyword{TVA}
\keyword{files}
\keyword{trigger}
\keyword{tva}
\keyword{validation}
