% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_anova_models.R
\name{cross_anova_models}
\alias{cross_anova_models}
\title{A function to help selecting the best model fitting according to the Chisquare between two models}
\usage{
cross_anova_models(models)
}
\arguments{
\item{models}{A list of lmer or glmer models under the form list(M1 = m1, M2 = m2, ..., Mn = mn). it is also possible to enter the data without names: c(m1, m2, m3) but this way of doing might be confusing when having many models.}
}
\description{
This function takes as input a list of models and returns a matrix containg all the models and showing the Chisquare difference between all models. The calculation is based on the anova() function, here simply performed for each model against each model
}
\keyword{GLMM}
\keyword{LMM}
\keyword{anova}
\keyword{glmer}
\keyword{lmer}
\keyword{models}
