% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.banner.R
\name{date.banner}
\alias{date.banner}
\title{A function to create a banner with the date inside}
\usage{
date.banner()
}
\description{
This function requiering the bannerCommenter package creates a banner with the date of the day inside. It helps keeping track of the changes in the code
}
\keyword{banner}
\keyword{date}
