% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.spaces.STEN.R
\name{delete.spaces.STEN}
\alias{delete.spaces.STEN}
\title{A function to prepare the data to be analyzed in the STEN software}
\usage{
delete.spaces.STEN(path)
}
\arguments{
\item{path}{Path to the files to rename}
}
\description{
This function removes the spaces and replace them by an underscore
}
\keyword{STEN}
\keyword{delete}
\keyword{spaces}
