% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_plots_LMM.R
\name{exp_plots_LMM}
\alias{exp_plots_LMM}
\title{A function to explore the data before undergoing statistics}
\usage{
exp_plots_LMM(data, DV, ID, class)
}
\arguments{
\item{data}{data frame containing the behavioral results}

\item{DV}{dependant variable under the form data$DV}

\item{ID}{ID variable containing the subjetcs names under the form data$ID}

\item{class}{variable containing the items or other second level variable under the form data$class}
}
\description{
This function draws some plots to explore the data before moving to the linear mixed models. The graphs do not represent all the needed information. However, it draws the main graph needed to perform a linear mixed model with the subject and items random effects.
}
\keyword{LMM}
\keyword{exploratory}
\keyword{lmer}
\keyword{mixed}
\keyword{models}
\keyword{plots}
