% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC_ranef.R
\name{ICC_ranef}
\alias{ICC_ranef}
\title{A function to extimate the ICC (intra-class correlation)}
\usage{
ICC_ranef(model)
}
\arguments{
\item{model}{A lmer or glmer model from the lme4 package}
}
\description{
This function allows, for a given lmer model, to estimate by how much a random effect explains the total variance of the model. Moreover, if the ICC is high, the difference inside the element of the class are important (large differences among the subjects for example). And if the ICC is small, it implies a coherence inside the class.
}
\keyword{ICC}
\keyword{glmer}
\keyword{lmer}
\keyword{variance}
