% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.epochs.ep.R
\name{import.epochs.ep}
\alias{import.epochs.ep}
\title{A function to import epochs in a data frame}
\usage{
import.epochs.ep(path, N_electrodes)
}
\arguments{
\item{path}{Path to the data in .ep format}

\item{N_electrodes}{Number of electrodes of the recording}
}
\description{
This function takes the path to the .ep files and give back a data framed following the tidyformat and under a list with each of the element being the different data frames corresponding to the data of one subject.
}
\keyword{epochs}
\keyword{import}
