% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_permuco.R
\name{import_data_permuco}
\alias{import_data_permuco}
\title{A function to prepare the signal's data for the brainperm function (permuco4brain)}
\usage{
import_data_permuco(path, N_electrodes, files_extension)
}
\arguments{
\item{path}{path to the folder containing the averaged signal per subjects and condition (tip: name the files wisely)}

\item{N_electrodes}{Number of electrodes included in the montage: either 128 or 64}

\item{files_extension}{Format of the files needed to be imported. Supported format are ".ep", ".txt", ".asc".}
}
\description{
This function build a list containing each individual's average signal and transforms it into a 3D matrix to be analyzed with the brainperm() function from the permuco4brain package. For this function to work you will need to install the readr package and the abind package
}
\keyword{import}
\keyword{mass}
\keyword{permuco}
\keyword{univariate}
