% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stats_Cartool.R
\name{import_stats_Cartool}
\alias{import_stats_Cartool}
\title{A function to import the microstates statistics from Cartool to R}
\usage{
import_stats_Cartool(path)
}
\arguments{
\item{path}{a character vector containing the path to the directory where the .seg and the statistics are stored.}
}
\description{
This function allows the user to directly import the microstates statistics from Cartool to R. Cartool gives one .csv file per within group comparison. This function takes the path of the folder containing the results, imports only the relevant .csv files and join them in one data frame, ready to analyse it.
}
\keyword{Cartool}
\keyword{import}
\keyword{microstates}
\keyword{statistics}
