% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stats_Ragu.R
\name{import_stats_Ragu}
\alias{import_stats_Ragu}
\title{A function to import stats data from Ragu into R}
\usage{
import_stats_Ragu(path, n_maps, n_conditions)
}
\arguments{
\item{path}{A path (e.g. "C:\\...") where the Ragu stats Excel file is located on your computer}

\item{n_maps}{Number of maps given by Ragu for the segmentation, i.e. the number of Excel sheets named "Class_n"}

\item{n_conditions}{Number of between subjects conditions defined in the Ragu settings}
}
\description{
This function turns the .xlsx exports data files from Ragu into a data frame compatible for
data analysis in R under the tidy format. To use the function attribute it to an element, just like
the read_excel function
}
\keyword{Ragu}
\keyword{microstates}
