% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fitting.R
\name{mod_fitting}
\alias{mod_fitting}
\title{A function to get the fitting information for a list of lmer of glmer models}
\usage{
mod_fitting(models)
}
\arguments{
\item{models}{A list of lmer or glmer models under the form list(M1 = m1, M2 = m2, ..., Mn = mn). it is also possible to enter the data without names: c(m1, m2, m3) but this way of doing might be confusing when having many models.}
}
\description{
This function takes as input a list of models and returns the AIC, BIC, log likelihood, deviance and residual degrees of freedom of each one of the lmer and glmer models. When REML is set to TRUE, the functions returns for this model the cAIC from the cAIC4 pkg. The models are then ordered according to the smallest AIC or cAIC.
}
\keyword{AIC}
\keyword{BIC}
\keyword{GLMM}
\keyword{LMM}
\keyword{deviance}
\keyword{df_resid}
\keyword{glmer}
\keyword{lmer}
\keyword{logLik}
