% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_word.R
\name{number_to_word}
\alias{number_to_word}
\title{A function to convert number into words for RMarkdown}
\usage{
number_to_word(number, cap = FALSE)
}
\arguments{
\item{number}{Integer to convert into word}

\item{cap}{If TRUE, the first letter of the word will be a capital. Default is FALSE}
}
\description{
It is usually conventional that integers below 10 should be written in words instead of numbers. In RMarkdown, when writing results, the the output of the results, the user can put the numbers in the present function to convert it automatically into text. If the result to display is 2, then number_toword(2) will return "two".
}
