% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{plot.weights}
\alias{plot.weights}
\title{Title}
\usage{
\method{plot}{weights}(dataset, model)
}
\arguments{
\item{dataset}{the dataset frame containing robust weights}

\item{model}{an object of class rlmerMod}
}
\value{
exploratory graphs, DV as a function of IV and weights
}
\description{
Title
}
