% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seg_file.R
\name{read_seg_file}
\alias{read_seg_file}
\title{A function to import microstates seg file from Cartool}
\usage{
read_seg_file(path)
}
\arguments{
\item{path}{path to the .seg file. e.g. "c:/Users/.../segfile.seg"}
}
\description{
This function allows the user to open a .seg file from Cartool (usually
obtained after a segmentation into microstates) in R. The function rename
the columns to put them in a freidly format.
}
\keyword{Cartool,}
\keyword{import}
\keyword{microstates,}
\keyword{segmentation,}
