% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_weights.R
\name{robust_weights}
\alias{robust_weights}
\title{add the random effects and observation weights to the dataframe}
\usage{
robust_weights(model_tested)
}
\arguments{
\item{model_tested}{an object of class lmerMod or rlmerMod to define the model to test}
}
\value{
the dataframe with robust weights
}
\description{
add the random effects and observation weights to the dataframe
}
