% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{A function to round up entire dataframe}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{a dataframe under the data frame format or a Tibble}

\item{digits}{number of decimal wanted}
}
\description{
This function enables tue user to round up all values in a dataframe. The functions isolate double variables and round them up.
}
\keyword{data}
\keyword{frame}
\keyword{round}
