% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_sd.R
\name{screen_n_sd}
\alias{screen_n_sd}
\title{Function Screen_sd}
\usage{
screen_n_sd(RT_var, min = 1, max = 4.5, step = 0.5, data)
}
\arguments{
\item{RT_var}{RT variable under the format data$variable.}

\item{min}{Lowest value of the range of SD to explore. Default is 1}

\item{max}{Highest value of the range of SD to explore. Default is 4.5}

\item{step}{Step from which to increase the SD to screen. Default is 0.5. For example with default values : 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5.}

\item{data}{Dataframe containing the data. Same as the element before the $ sign in the RT_var parameter.}
}
\description{
This function allows to select how many standard deviations above and under the mean to remove for data cleaning.
}
\keyword{SD}
\keyword{deviation}
\keyword{distribution}
\keyword{screening}
\keyword{standard}
