% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_error.R
\name{std_error}
\alias{std_error}
\title{Function to estimate the standard error}
\usage{
std_error(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numerical array}

\item{na.rm}{argument to set if NA are present in the array and exclude them}
}
\description{
This function estimates the standard error of an array. The formula is the square root of the variance / the square root of the sample size.
}
