% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.subj.group.R
\name{clean.subj.group}
\alias{clean.subj.group}
\title{A function to remove outlier subjects}
\usage{
clean.subj.group(group.var, subj.var, RT.var, N.sd, data)
}
\arguments{
\item{group.var}{Groupping variable for example age group}
}
\description{
This function examines the data and gives back which subjects are to be removed because of mean RT above or under a certain number of standard deviations
}
\keyword{RT}
\keyword{SD}
\keyword{clean}
\keyword{deviations}
\keyword{group}
\keyword{standard}
\keyword{subject}
