% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.subj.group.R
\name{clean.subj.group}
\alias{clean.subj.group}
\title{A function to remove outlier subjects}
\usage{
clean.subj.group(group.var, subj.var, RT.var, N.sd, data)
}
\arguments{
\item{group.var}{Groupping variable for example age group. Use the raw name of the column e.g. group instead of data$group without "" either}

\item{subj.var}{Subject variable. Use the raw name of the column e.g. subjects instead of data$subjects without "" either}

\item{RT.var}{Variable containing the reaction times. Use the raw name of the column e.g. RT instead of data$RT without "" either}

\item{N.sd}{Number of standard deviation above and under the mean where to put the rejection boundary.}

\item{data}{Name of the dataframe containing the experimental data.}
}
\description{
This function examines the data and gives back which subjects are to be removed because of mean RT above or under a certain number of standard deviations
}
\keyword{RT}
\keyword{SD}
\keyword{clean}
\keyword{deviations}
\keyword{group}
\keyword{standard}
\keyword{subject}
